Optimized speed and decreased memory usage
Since we just need the closest pair of numbers, we don't need to store every multiple. In fact, we only need to store one pair of multiples.
We can optimistically start at the closest pair, the square root, and widen the range by 1 each time. When we reach a pair of numbers that multiply to the target, then we know we are done.

Multiples of large numbers can now be found without running out of memory.

Also removed some extra variable declarations.